services/inference_client/pyproject.toml
[project]
name = "inference-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.9"
dependencies = [
    "comet-ml>=3.49.10",
    "geopandas>=1.0.1",
    "holidays>=0.72",
    "joblib>=1.5.0",
    "loguru>=0.7.3",
    "openmeteo-requests>=1.4.0",
    "packaging>=25.0",
    "pandas>=2.2.3",
    "plotly>=6.1.1",
    "pydantic-settings>=2.9.1",
    "requests-cache>=1.2.1",
    "retry-requests>=2.0.0",
    "streamlit>=1.22.0",
    "streamlit-bokeh-events>=0.1.2",
    "xgboost>=3.0.1",
]

[dependency-groups]
dev = []
